backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.10.0
sqlalchemy>=2.0.30
requests>=2.32.0
dash>=2.18.0
plotly>=5.24.0
pandas>=2.1.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.9
aiofiles>=24.1.0
httpx>=0.27.0
a2wsgi
openai
openpyxl>=3.1.0
reportlab>=4.2.0
